Move auto docs deeper in documentation
