+++ pgdriver/auto/auto.go
@@ -1,4 +1,22 @@
-package pgdriverauto
+// Package auto allows one to setup rebecca/pgdriver in a convenient way from
+// environment variables.
+//
+// To start using auto use this nameless import:
+//
+// 				import _ "github.com/waterlink/rebecca/pgdriver/auto"
+//
+// Empty import will fetch connection options from respective environment
+// variables:
+//
+// - `REBECCA_PG_USER`
+// - `REBECCA_PG_PASS`
+// - `REBECCA_PG_HOST`
+// - `REBECCA_PG_PORT`
+// - `REBECCA_PG_DATABASE`
+// - `REBECCA_PG_SSLMODE`
+//
+// or, if present: `REBECCA_PG_URL`
+package auto
 
 import (
 	"fmt"
